refactor(orchestration): decompose ExecutionStrategyHandler into Strategy pattern (GH #6830)
Split 317-LOC god class into focused per-strategy files:
- _base.py: BaseExecutionStrategy ABC + _safe_execute + _is_required_failure
- _sequential.py: SequentialStrategy with _wait_for_dependencies
- _parallel.py: ParallelStrategy with deadlock detection
- _pipeline.py: PipelineStrategy with stage-based execution
- _collaborative.py: CollaborativeStrategy with Redis pubsub
- _adaptive.py: AdaptiveStrategy with strategy switching based on metrics
- __init__.py: ExecutionStrategyHandler thin dispatcher (≤100 LOC)

Backward compat preserved: ExecutionStrategyHandler import path unchanged.

Acceptance criteria (GH #6830):
- ✅ dispatcher ≤100 LOC
- ✅ one file per strategy ≤200 LOC each
- ✅ each strategy has focused responsibility
- ✅ import path preserved for callers
