--- autobot-backend/enhanced_orchestration/execution_strategies.py
@@ -1,296 +1,0 @@
-# AutoBot - AI-Powered Automation Platform
-# Copyright (c) 2025 mrveiss
-# Author: mrveiss
-"""
-Execution Strategies Module
-
-Issue #381: Extracted from enhanced_multi_agent_orchestrator.py god class refactoring.
-Contains execution strategy implementations for workflow orchestration.
-"""
-
-import asyncio
-from typing import Any, Dict, Tuple
-
-from autobot_shared.logging_manager import get_logger
-from constants.threshold_constants import TimingConstants
-
-from .types import AgentTask, ExecutionStrategy, WorkflowDependencies, WorkflowPlan
-
-logger = get_logger(__name__)
-
-
-class ExecutionStrategyHandler:
-    """Handles different execution strategies for workflow plans."""
-
-    def __init__(
-        self,
-        max_parallel_tasks: int,
-        resource_semaphore: asyncio.Semaphore,
-        deps: WorkflowDependencies,
-    ) -> None:
-        self.max_parallel_tasks = max_parallel_tasks
-        self.resource_semaphore = resource_semaphore
-        self._execute_single_task = deps.execute_single_task
-        self._topological_sort_tasks = deps.topological_sort_tasks
-        self._dependencies_met = deps.dependencies_met
-        self._group_pipeline_stages = deps.group_pipeline_stages
-        self._enhance_task_for_collaboration = deps.enhance_task_for_collaboration
-        self._coordinate_collaboration = deps.coordinate_collaboration
-        self.coordination_prefix = "autobot:orchestrator:coord:"
-
-    async def execute_by_strategy(self, plan: WorkflowPlan) -> Dict[str, Any]:
-        """Execute workflow based on its strategy (Issue #315: extracted)."""
-        strategy_handlers = {
-            ExecutionStrategy.SEQUENTIAL: self.execute_sequential,
-            ExecutionStrategy.PARALLEL: self.execute_parallel,
-            ExecutionStrategy.PIPELINE: self.execute_pipeline,
-            ExecutionStrategy.COLLABORATIVE: self.execute_collaborative,
-        }
-        handler = strategy_handlers.get(plan.strategy, self.execute_adaptive)
-        return await handler(plan)
-
-    async def execute_sequential(self, plan: WorkflowPlan) -> Dict[str, Any]:
-        """Execute tasks sequentially"""
-        results = {}
-
-        # Sort tasks by dependencies and priority
-        sorted_tasks = self._topological_sort_tasks(plan.tasks, plan.dependencies_graph)
-
-        for task in sorted_tasks:
-            logger.info("Executing task %s (%s)", task.task_id, task.agent_type)
-
-            try:
-                await self._wait_for_dependencies(task, results)
-            except RuntimeError as exc:
-                result = task.to_failed_result(str(exc))
-                results[task.task_id] = result
-                if self._is_required_failure(task, result):
-                    logger.error("Required task %s blocked by failed dep, stopping", task.task_id)
-                    break
-                continue
-
-            result = await self._safe_execute(task, results)
-            results[task.task_id] = result
-
-            if self._is_required_failure(task, result):
-                logger.error("Required task %s failed, stopping workflow", task.task_id)
-                break
-
-        return results
-
-    async def execute_parallel(self, plan: WorkflowPlan) -> Dict[str, Any]:
-        """Execute independent tasks in parallel"""
-        results = {}
-        pending_tasks = list(plan.tasks)
-        running_tasks = []
-
-        while pending_tasks or running_tasks:
-            # Start tasks that have their dependencies met
-            ready_tasks = []
-            for task in pending_tasks[:]:
-                if self._dependencies_met(task, results):
-                    ready_tasks.append(task)
-                    pending_tasks.remove(task)
-
-            # Start ready tasks (respecting resource limits)
-            for task in ready_tasks:
-                if len(running_tasks) < self.max_parallel_tasks:
-                    logger.info("Starting parallel task %s", task.task_id)
-                    task_future = asyncio.create_task(self._safe_execute(task, results))
-                    running_tasks.append((task, task_future))
-
-            # Wait for any task to complete
-            if running_tasks:
-                done, _ = await asyncio.wait(
-                    [future for _, future in running_tasks],
-                    return_when=asyncio.FIRST_COMPLETED,
-                )
-
-                # Process completed tasks
-                for task, future in running_tasks[:]:
-                    if future in done:
-                        result = await future
-                        results[task.task_id] = result
-                        running_tasks.remove((task, future))
-                        logger.info("Completed parallel task %s", task.task_id)
-            elif pending_tasks:
-                # No running tasks but pending remain — dependency deadlock (#6420)
-                logger.error(
-                    "Dependency deadlock: %d tasks unresolvable, failing them",
-                    len(pending_tasks),
-                )
-                for task in pending_tasks:
-                    results[task.task_id] = task.to_failed_result("Dependency deadlock")
-                break
-
-        return results
-
-    async def execute_pipeline(self, plan: WorkflowPlan) -> Dict[str, Any]:
-        """Execute tasks in pipeline mode where outputs feed into next inputs"""
-        results = {}
-
-        # Group tasks into pipeline stages
-        stages = self._group_pipeline_stages(plan.tasks, plan.dependencies_graph)
-
-        pipeline_data = {}
-        for stage_num, stage_tasks in enumerate(stages):
-            logger.info("Executing pipeline stage %d/%d", stage_num + 1, len(stages))
-
-            stage_results = await asyncio.gather(
-                *[self._safe_execute(task, {**results, **pipeline_data}) for task in stage_tasks]
-            )
-
-            stage_failed = False
-            for task, result in zip(stage_tasks, stage_results):
-                results[task.task_id] = result
-                if result.get("status") == "completed" and "output" in result:
-                    pipeline_data.update(result["output"])
-                if self._is_required_failure(task, result):
-                    stage_failed = True
-
-            if stage_failed:
-                logger.error("Pipeline stage %d has required failures, stopping", stage_num + 1)
-                break
-
-        return results
-
-    async def execute_collaborative(self, plan: WorkflowPlan) -> Dict[str, Any]:
-        """Execute tasks collaboratively with inter-agent communication"""
-        results = {}
-
-        # Create collaboration channels
-        collab_channel = f"{self.coordination_prefix}collab:{plan.plan_id}"
-
-        # Start collaboration coordinator
-        coordinator_task = asyncio.create_task(self._coordinate_collaboration(collab_channel))
-
-        # Execute tasks with collaboration
-        task_futures = []
-        for task in plan.tasks:
-            enhanced_task = self._enhance_task_for_collaboration(task, collab_channel)
-            future = asyncio.create_task(self._safe_execute(enhanced_task, results))
-            task_futures.append((task, future))
-
-        # Wait for all tasks
-        for task, future in task_futures:
-            result = await future
-            results[task.task_id] = result
-
-        # Stop coordinator and wait for its cleanup (finally/unsubscribe) to complete (#6431)
-        coordinator_task.cancel()
-        try:
-            await coordinator_task
-        except asyncio.CancelledError:
-            pass
-
-        return results
-
-    def _adapt_strategy(
-        self, progress_ratio: float, failure_ratio: float, current: ExecutionStrategy
-    ) -> ExecutionStrategy:
-        """Adapt execution strategy based on progress metrics."""
-        if failure_ratio > 0.3:
-            logger.info("Adapting to SEQUENTIAL due to high failure rate")
-            return ExecutionStrategy.SEQUENTIAL
-
-        if progress_ratio > 0.7 and failure_ratio < 0.1:
-            logger.info("Adapting to PARALLEL due to good progress")
-            return ExecutionStrategy.PARALLEL
-
-        return current
-
-    def _is_required_failure(self, task: AgentTask, result: dict) -> bool:
-        return result.get("status") == "failed" and not task.metadata.get("optional", False)
-
-    async def _safe_execute(self, task: AgentTask, ctx: Dict[str, Any]) -> dict:
-        """Run _execute_single_task and convert any unhandled Exception to a failed result (#6459).
-
-        CancelledError is re-raised so cooperative cancellation still works.
-        """
-        try:
-            return await self._execute_single_task(task, ctx)
-        except asyncio.CancelledError:
-            raise
-        except Exception as exc:
-            logger.exception("Task %s raised during execution", task.task_id)
-            return task.to_failed_result(repr(exc))
-
-    async def _execute_parallel_batch(self, pending_tasks: list, results: Dict[str, Any]) -> Tuple[int, int]:
-        """Execute tasks in parallel batch."""
-        batch_size = min(self.max_parallel_tasks, len(pending_tasks))
-        batch_tasks = pending_tasks[:batch_size]
-        ready_tasks = [t for t in batch_tasks if self._dependencies_met(t, results)]
-
-        batch_results = await asyncio.gather(*[self._safe_execute(task, results) for task in ready_tasks])
-
-        completed, failed = 0, 0
-        for task, result in zip(ready_tasks, batch_results):
-            results[task.task_id] = result
-            pending_tasks.remove(task)
-            if result.get("status") == "completed":
-                completed += 1
-            else:
-                failed += 1
-
-        return completed, failed
-
-    async def _execute_sequential_step(self, pending_tasks: list, results: Dict[str, Any]) -> Tuple[int, int]:
-        """Execute one sequential task step."""
-        for task in pending_tasks[:]:
-            if not self._dependencies_met(task, results):
-                continue
-
-            result = await self._safe_execute(task, results)
-            results[task.task_id] = result
-            pending_tasks.remove(task)
-
-            if result.get("status") == "completed":
-                return 1, 0
-            return 0, 1
-
-        return 0, 0
-
-    async def execute_adaptive(self, plan: WorkflowPlan) -> Dict[str, Any]:
-        """Execute with adaptive strategy (Issue #315 - refactored)."""
-        results = {}
-        current_strategy = plan.strategy
-        completed_tasks, failed_tasks = 0, 0
-        pending_tasks = list(plan.tasks)
-
-        while pending_tasks:
-            progress_ratio = completed_tasks / len(plan.tasks)
-            failure_ratio = failed_tasks / max(completed_tasks, 1)
-            current_strategy = self._adapt_strategy(progress_ratio, failure_ratio, current_strategy)
-
-            if current_strategy == ExecutionStrategy.PARALLEL:
-                c, f = await self._execute_parallel_batch(pending_tasks, results)
-            else:
-                c, f = await self._execute_sequential_step(pending_tasks, results)
-
-            if c == 0 and f == 0 and pending_tasks:
-                # No progress made — dependency deadlock (#6429)
-                logger.error(
-                    "Dependency deadlock in adaptive: %d tasks unresolvable, failing them",
-                    len(pending_tasks),
-                )
-                for task in pending_tasks:
-                    results[task.task_id] = task.to_failed_result("Dependency deadlock")
-                break
-
-            completed_tasks += c
-            failed_tasks += f
-
-        return results
-
-    async def _wait_for_dependencies(self, task: AgentTask, results: Dict[str, Any]) -> None:
-        """Wait for task dependencies to complete, or raise if a dep is in any terminal non-completed state."""
-        while not self._dependencies_met(task, results):
-            terminal_deps = [
-                dep for dep in task.dependencies if dep in results and results[dep].get("status") != "completed"
-            ]
-            if terminal_deps:
-                status = results[terminal_deps[0]].get("status")
-                raise RuntimeError(
-                    f"Task {task.task_id} cannot run: dependency {terminal_deps[0]}" f" is terminal (status={status})"
-                )
-            await asyncio.sleep(TimingConstants.SHORT_DELAY)
